+++ clang-tools-extra/test/clang-tidy/modernize-use-nodiscard.h
@@ -1,0 +1,5 @@
+
+#define MUST_USE_RESULT __attribute__((warn_unused_result))
+#define NO_DISCARD [[nodiscard]]
+#define NO_RETURN [[noreturn]]
+